Convert System To Package Structure Format
MAJOR CHANGES
1) Package File Structure
Switch to using the package structure format. Each api client has been
broken into it's own module (imgur, reddit, twitter) as well as the
database and class that links them together. The primary file to call
to run the whole thing is now `main.py`.

2) Database Structure
The database structure has also been changed to make the key names more
clear and verbose. It's also been seperated into a deeper document so
there are top level keys for the comic, twitter, imgur, and reddit.
Those then have related values under them (twitter user name is under
the twitter key). There is also a new meta table added that keeps info
about the comic as a whole. Info like the current number, imgur album
info, subreddit to post to, twitter user info, ect.

Because the new format is incompatible with the old one a convert tool
is included in this release (and will be removed in future ones). It's
advisable to make a backup of the database before running the conversion.
The tool will create the `meta` table. Number may be improperly added
and will need to be manually set. It then converts all the entries to
the best of it's abilities.

MINOR CHANGES
1) Disable Reddit inbox replies
Reddit submissions have disabled inbox replies.

2) Switch to database tracking number
Database now tracks the current comic number in the meta table
instead of an external file

3) Better Logging
Added some more clear logging. Stoped using logging.warn which has been
deprecated and switch to warning.

BUG FIXES
1) Reverse recent tweet status list
Tweet statuses are recieved from newest to oldest. Used to number them in
that order to which is backwards. Now reverse the list in the
TwitterApiClient when it gets the list.

2) Retries check if there are attempted posts
When there are no new posts to make, `upload()` returns `None` otherwise
it returns the list of reddit submission urls. Previously the retry loop
was checking if the returned value was falsy (None or empty list would be
valid). Now it checks that it was not None and has a len of 0 to retry
(and still less than the max attemps, that hasn't changed).

3) Update Python Requirements
- praw: 7.1.4 -> 7.2.0
- tinydb: 4.3.0 -> 4.4.0
- python-dotenv: 0.15.0 -> 0.16.0

FORMATING CHANGES
1) Add Docstrings
Uses reStructuredText format (there may be some google style leftover).

2) Added type hints
Class, function, and method signartures have type hints added for their
arguments and return values. Variables hints will be added in future
updates.

3) Switch string formatting
Removed most `str.format()` calls and replaced with `f"string"` style.
Logging swiched to the `"%s", str` style to reduce on unneeded
computations if the level isn't logged.

4) Formatting with Black, Pylint, and Isort
Now using black, pylint, and isort to help keep things as close to pep-8
as possible and making things uniform across the codebase. There will be
some personal quirks here and there, but (at least with pylint) those
get maked with a `#pylint: disable=warning` comment.